refactor: Reduce cycle count from 19 to 8
