+++ use-core/src/main/java/org/tzi/use/uml/ocl/extension/ExtensionOperation.java
@@ -7,8 +7,6 @@ import org.tzi.use.uml.ocl.type.Type;
 import org.tzi.use.uml.ocl.value.UndefinedValue;
 import org.tzi.use.uml.ocl.value.Value;
 import org.tzi.use.util.Log;
-import org.tzi.use.util.rubyintegration.RubyHelper;
-
 import javax.script.ScriptContext;
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
